Renamed folders (mongo refactoring)
